# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HR Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentBody"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      responses:
        "204":
          description: Deleted

  # ── EMPLOYEES ──────────────────────────────────────────────────────────────
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # ── USERS (admin only) ───────────────────────────────────────────────────
...
  /demands:
    get:
      operationId: listDemands
      tags: [demands]
      summary: List shift demands for a given month
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of demands with decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DemandWithDecision"
    post:
      operationId: createDemand
      tags: [demands]
      summary: Create or update a shift demand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDemandBody"
      responses:
        "201":
          description: Created/updated demand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "400":
          description: Validation error (past month, locked month, locked day)
  /demands/{id}:
    delete:
      operationId: deleteDemand
      tags: [demands]
      summary: Delete a demand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /demands/{id}/decision:
    patch:
      operationId: decideDemand
      tags: [demands]
      summary: Approve or reject a demand (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandDecisionBody"
      responses:
        "200":
          description: Updated demand and decision record
          content:
            application/json:
              schema:
                type: object
                properties:
                  demand:
                    $ref: "#/components/schemas/Demand"
                  decision:
                    $ref: "#/components/schemas/DemandDecision"

  # ── MAIL SETTINGS (admin only) ───────────────────────────────────────────
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        employeeId:
          type: integer
          nullable: true
      required: [id, username, role]

    # USERS
...
    Demand:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        year:
          type: integer
        month:
          type: integer
        day:
          type: integer
        demandCode:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        notifiedAt:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, employeeId, year, month, day, demandCode, status]
    DemandDecision:
      type: object
      properties:
        id:
          type: integer
        demandId:
          type: integer
        adminId:
          type: integer
          nullable: true
        decision:
          type: string
          enum: [approved, rejected]
        notifiedAt:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, demandId, decision]
    DemandWithDecision:
      allOf:
        - $ref: "#/components/schemas/Demand"
        - type: object
          properties:
            decision:
              $ref: "#/components/schemas/DemandDecision"
              nullable: true
    CreateDemandBody:
      type: object
      properties:
        employeeId:
          type: integer
        year:
          type: integer
        month:
          type: integer
        day:
          type: integer
        demandCode:
          type: string
      required: [employeeId, year, month, day, demandCode]
    DemandDecisionBody:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
      required: [decision]

    # MAIL SETTINGS
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, order, createdAt, updatedAt]
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer
      required: [name]
    UpdateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer

    # EMPLOYEE
...